Extract class Main from bin/ts-java.ts to lib/ts-java-main.ts
